add and refactor io tests
